@@ -1,7 +1,7 @@
 # encoding: utf-8
 
-ActiveAdmin::Dashboards.build do
-
+ActiveAdmin.register_page "Dashboard" do
+  
   # Define your dashboard sections here. Each block will be
   # rendered on the dashboard in the context of the view. So just
   # return the content which you would like to display.
@@ -9,69 +9,69 @@ ActiveAdmin::Dashboards.build do
   # == Simple Dashboard Section
   # Here is an example of a simple dashboard section
   #
-   section I18n.t('active_admin.dashboards.article_section'), priority: 1, :if => proc{can?(:update, Goldencobra::Article)} do
-    table do
-      tr do
-        [I18n.t("activerecord.attributes.goldencobra/article.title"), I18n.t("activerecord.attributes.goldencobra/article.created_at"), ""].each do |sa|
-          th sa
-        end
-      end
+   # section I18n.t('active_admin.dashboards.article_section'), priority: 1, :if => proc{can?(:update, Goldencobra::Article)} do
+   #  table do
+   #    tr do
+   #      [I18n.t("activerecord.attributes.goldencobra/article.title"), I18n.t("activerecord.attributes.goldencobra/article.created_at"), ""].each do |sa|
+   #        th sa
+   #      end
+   #    end
 
-      Goldencobra::Article.recent(5).collect do |article|
-        tr do
-          td article.title
-          td l(article.created_at, format: :short)
-          result = link_to(t(:view), article.public_url, :class => "member_link edit_link view", :title => I18n.t('active_admin.dashboards.title1'), :target => "_blank")
-          result += link_to(t(:edit), admin_article_path(article), :class => "member_link edit_link edit", :title => I18n.t('active_admin.dashboards.title2'))
-          result += link_to(t(:new_subarticle), new_admin_article_path(:parent => article), :class => "member_link edit_link new_subarticle", :title => I18n.t('active_admin.dashboards.title3'))
-          td result
-        end
-      end
-    end
+   #    Goldencobra::Article.recent(5).collect do |article|
+   #      tr do
+   #        td article.title
+   #        td l(article.created_at, format: :short)
+   #        result = link_to(t(:view), article.public_url, :class => "member_link edit_link view", :title => I18n.t('active_admin.dashboards.title1'), :target => "_blank")
+   #        result += link_to(t(:edit), admin_article_path(article), :class => "member_link edit_link edit", :title => I18n.t('active_admin.dashboards.title2'))
+   #        result += link_to(t(:new_subarticle), new_admin_article_path(:parent => article), :class => "member_link edit_link new_subarticle", :title => I18n.t('active_admin.dashboards.title3'))
+   #        td result
+   #      end
+   #    end
+   #  end
 
-    table do
-      tr do
-        td link_to(I18n.t('active_admin.dashboards.new_link'), admin_article_path("new"))
-      end
-    end
-   end
+   #  table do
+   #    tr do
+   #      td link_to(I18n.t('active_admin.dashboards.new_link'), admin_article_path("new"))
+   #    end
+   #  end
+   # end
 
-   section I18n.t('active_admin.dashboards.widget_section'), priority: 2, :if => proc{can?(:update, Goldencobra::Widget)} do
-    table do
-      tr do
-        [I18n.t("activerecord.attributes.goldencobra/widget.title"), I18n.t("activerecord.attributes.goldencobra/widget.created_at"), ""].each do |sa|
-          th sa
-        end
-      end
+   # section I18n.t('active_admin.dashboards.widget_section'), priority: 2, :if => proc{can?(:update, Goldencobra::Widget)} do
+   #  table do
+   #    tr do
+   #      [I18n.t("activerecord.attributes.goldencobra/widget.title"), I18n.t("activerecord.attributes.goldencobra/widget.created_at"), ""].each do |sa|
+   #        th sa
+   #      end
+   #    end
 
-      Goldencobra::Widget.recent(5).collect do |widget|
-        tr do
-          td widget.title
-          td l(widget.created_at, format: :short)
-          td link_to(t(:edit), admin_widget_path(widget), :class => "member_link edit_link edit", :title => I18n.t('active_admin.dashboards.title4'))
-        end
-      end
-    end
-   end
+   #    Goldencobra::Widget.recent(5).collect do |widget|
+   #      tr do
+   #        td widget.title
+   #        td l(widget.created_at, format: :short)
+   #        td link_to(t(:edit), admin_widget_path(widget), :class => "member_link edit_link edit", :title => I18n.t('active_admin.dashboards.title4'))
+   #      end
+   #    end
+   #  end
+   # end
 
-    section I18n.t('active_admin.dashboards.vita_steps'), priority: 2, :if => proc{can?(:update, Goldencobra::Vita)} do
-        table do
-          tr do
-            ["Source", I18n.t("activerecord.attributes.goldencobra/widget.title"), "Description", ""].each do |sa|
-              th sa
-            end
-          end
+   #  section I18n.t('active_admin.dashboards.vita_steps'), priority: 2, :if => proc{can?(:update, Goldencobra::Vita)} do
+   #      table do
+   #        tr do
+   #          ["Source", I18n.t("activerecord.attributes.goldencobra/widget.title"), "Description", ""].each do |sa|
+   #            th sa
+   #          end
+   #        end
 
-          Goldencobra::Vita.where(status_cd: 2).last(5).each do |vita|
-            tr do
-              td "#{vita.loggable_type} ID:#{vita.loggable_id}"
-              td vita.title
-              td vita.description
-              td l(vita.created_at, format: :short)
-            end
-          end
-        end
-     end
+   #        Goldencobra::Vita.where(status_cd: 2).last(5).each do |vita|
+   #          tr do
+   #            td "#{vita.loggable_type} ID:#{vita.loggable_id}"
+   #            td vita.title
+   #            td vita.description
+   #            td l(vita.created_at, format: :short)
+   #          end
+   #        end
+   #      end
+   #   end
 
   # == Render Partial Section
   # The block is rendered within the context of the view, so you can
