@@ -8,7 +8,7 @@ ActiveAdmin.register Goldencobra::Widget, as: "Widget" do
   filter :id_name, :label => I18n.t('active_admin.widget.id')
   filter :sorter, :label => I18n.t('active_admin.widget.sorter')
 
-  scope I18n.t('active_admin.widget.scope1'), :scoped, :default => true
+  scope I18n.t('active_admin.widget.scope1'), :all, :default => true
   scope I18n.t('active_admin.widget.scope2'), :active
   scope I18n.t('active_admin.widget.scope3'), :inactive
   scope I18n.t('active_admin.widget.scope4'), :default
@@ -30,39 +30,42 @@ ActiveAdmin.register Goldencobra::Widget, as: "Widget" do
     f.inputs I18n.t('active_admin.widget.layout_web'), :class => "foldable inputs" do
       f.input :content, :label => I18n.t('active_admin.widget.label_web'), :hint => I18n.t('active_admin.widget.hint_web')
     end
-    f.inputs I18n.t('active_admin.widget.layout_mobile'), :class => "foldable inputs closed" do
-      f.input :mobile_content, :label => I18n.t('active_admin.widget.label_mobile'), :hint => I18n.t('active_admin.widget.hint_mobile')
-    end
+    # Wird so gut wie nie benutzt => erstmal ausgeblendet:
+    # f.inputs I18n.t('active_admin.widget.layout_mobile'), :class => "foldable inputs closed" do
+    #   f.input :mobile_content, :label => I18n.t('active_admin.widget.label_mobile'), :hint => I18n.t('active_admin.widget.hint_mobile')
+    # end
     f.inputs I18n.t('active_admin.widget.info'), :class => "foldable inputs closed"  do
       f.input :sorter, :label => I18n.t('active_admin.widget.label_info1'), :hint => I18n.t('active_admin.widget.label_hint1')
       f.input :css_name, :label => I18n.t('active_admin.widget.label_info2'), :hint => I18n.t('active_admin.widget.label_hint2')
       f.input :id_name, :label => I18n.t('active_admin.widget.label_info3'), :hint => I18n.t('active_admin.widget.label_hint3')
-      f.input :teaser
-      f.input :description, :label => I18n.t('active_admin.widget.label_info4'), :hint => I18n.t('active_admin.widget.label_hint4')
+      # Wird so gut wie nie benutzt => erstmal ausgeblendet:
+      #f.input :teaser
+      #f.input :description, :label => I18n.t('active_admin.widget.label_info4'), :hint => I18n.t('active_admin.widget.label_hint4')
     end
-    if Goldencobra::Setting.for_key("goldencobra.widgets.time_control") == "true"
-      f.inputs I18n.t('active_admin.widget.time'), :class => "foldable inputs closed" do
-        f.input :offline_time_active, :label => I18n.t('active_admin.widget.time_label1'), hint: I18n.t('active_admin.widget.time_hint1')
-        f.input :offline_date_start, :label => I18n.t('active_admin.widget.time_label2'), :hint => I18n.t('active_admin.widget.time_hint2')
-        f.input :offline_date_end, :label => I18n.t('active_admin.widget.time_label3'), :hint => I18n.t('active_admin.widget.time_hint3')
-        f.input :offline_day, :label => I18n.t('active_admin.widget.time_label4'), as: :check_boxes, collection: Goldencobra::Widget::OfflineDays
-        f.input :offline_time_start_mo, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_mo) }
-        f.input :offline_time_end_mo, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_mo) }
-        f.input :offline_time_start_tu, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_tu) }
-        f.input :offline_time_end_tu, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_tu) }
-        f.input :offline_time_start_we, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_we) }
-        f.input :offline_time_end_we, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_we) }
-        f.input :offline_time_start_th, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_th) }
-        f.input :offline_time_end_th, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_th) }
-        f.input :offline_time_start_fr, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_fr) }
-        f.input :offline_time_end_fr, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_fr) }
-        f.input :offline_time_start_sa, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_sa) }
-        f.input :offline_time_end_sa, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_sa) }
-        f.input :offline_time_start_su, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_su) }
-        f.input :offline_time_end_su, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_su) }
-        f.input :alternative_content, :label => "Alternativer Inhalt", hint: 'Dieser Inhalt wird angezeigt, wenn das Schnipsel offline ist, HTML möglich.'
-      end
-    end
+    # TODO Auslagern in ein eigenes Modul:
+    # if Goldencobra::Setting.for_key("goldencobra.widgets.time_control") == "true"
+    #   f.inputs I18n.t('active_admin.widget.time'), :class => "foldable inputs closed" do
+    #     f.input :offline_time_active, :label => I18n.t('active_admin.widget.time_label1'), hint: I18n.t('active_admin.widget.time_hint1')
+    #     f.input :offline_date_start, :label => I18n.t('active_admin.widget.time_label2'), :hint => I18n.t('active_admin.widget.time_hint2')
+    #     f.input :offline_date_end, :label => I18n.t('active_admin.widget.time_label3'), :hint => I18n.t('active_admin.widget.time_hint3')
+    #     f.input :offline_day, :label => I18n.t('active_admin.widget.time_label4'), as: :check_boxes, collection: Goldencobra::Widget::OfflineDays
+    #     f.input :offline_time_start_mo, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_mo) }
+    #     f.input :offline_time_end_mo, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_mo) }
+    #     f.input :offline_time_start_tu, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_tu) }
+    #     f.input :offline_time_end_tu, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_tu) }
+    #     f.input :offline_time_start_we, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_we) }
+    #     f.input :offline_time_end_we, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_we) }
+    #     f.input :offline_time_start_th, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_th) }
+    #     f.input :offline_time_end_th, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_th) }
+    #     f.input :offline_time_start_fr, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_fr) }
+    #     f.input :offline_time_end_fr, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_fr) }
+    #     f.input :offline_time_start_sa, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_sa) }
+    #     f.input :offline_time_end_sa, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_sa) }
+    #     f.input :offline_time_start_su, as: :string, placeholder: I18n.t(:offline_time_start_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_start_su) }
+    #     f.input :offline_time_end_su, as: :string, placeholder:  I18n.t(:offline_time_end_hint, scope: [:activerecord, :attributes, 'goldencobra/widget']), input_html: { value: (f.object.get_offline_time_end_su) }
+    #     f.input :alternative_content, :label => "Alternativer Inhalt", hint: 'Dieser Inhalt wird angezeigt, wenn das Schnipsel offline ist, HTML möglich.'
+    #   end
+    # end
     f.inputs I18n.t('active_admin.widget.access_rights'), :class => "foldable closed inputs" do
       f.has_many :permissions do |p|
         p.input :role, :include_blank => "Alle"
@@ -71,7 +74,7 @@ ActiveAdmin.register Goldencobra::Widget, as: "Widget" do
       end
     end
     f.inputs I18n.t('active_admin.widget.article') do
-      f.input :articles, :label => I18n.t('active_admin.widget.article_label'), :hint => I18n.t('active_admin.widget.article_hint'), :as => :select, :collection => Goldencobra::Article.find(:all, :order => "title ASC"), :input_html => { :class => 'chzn-select', "data-placeholder" => I18n.t('active_admin.widget.article_placeholder') }
+      f.input :articles, :label => I18n.t('active_admin.widget.article_label'), :hint => I18n.t('active_admin.widget.article_hint'), :as => :select, :collection => Goldencobra::Article.order("title ASC"), :input_html => { :class => 'chzn-select', "data-placeholder" => I18n.t('active_admin.widget.article_placeholder') }
     end
     f.actions
   end
@@ -179,7 +182,7 @@ ActiveAdmin.register Goldencobra::Widget, as: "Widget" do
   controller do
     def show
       show! do |format|
-         format.html { redirect_to edit_admin_widget_path(@widget), :flash => flash }
+         format.html { redirect_to edit_admin_widget_path(@widget) }
       end
     end
   end
