@@ -25,11 +25,11 @@ module Goldencobra
 
     validates_presence_of :article_id, :content, :commentator_id
 
-    scope :approved, where(:approved => true)
-    scope :not_approved, where(:approved => false)
-    scope :active, where(:active => true)
-    scope :reported, where(:reported => true)
-    scope :not_reported_and_active, where(reported: false, active: true)
+    scope :approved, -> { where(:approved => true) }
+    scope :not_approved, -> { where(:approved => false) }
+    scope :active, -> { where(:active => true) }
+    scope :reported, -> { where(:reported => true) }
+    scope :not_reported_and_active, -> { where(reported: false, active: true) }
 
     def title
       [self.article.title,self.content[0..20]].join(" - ")
