@@ -33,7 +33,7 @@ module Goldencobra
 
     has_many  :article_widgets
     has_many  :articles, :through => :article_widgets
-    has_many  :permissions, :class_name => Goldencobra::Permission, :foreign_key => "subject_id", :conditions => {:subject_class => "Goldencobra::Widget"}
+    has_many  :permissions, -> { where subject_class: "Goldencobra::Widget" }, class_name: Goldencobra::Permission, foreign_key: "subject_id"
 
     accepts_nested_attributes_for :permissions, :allow_destroy => true
 
@@ -49,10 +49,10 @@ module Goldencobra
 
     ImportDataFunctions = []
 
-    scope :active, where(:active => true).order(:sorter)
-    scope :inactive, where(:active => false).order(:sorter)
-    scope :default, where(:default => true).order(:sorter)
-    scope :not_default, where(:default => false).order(:sorter)
+    scope :active, -> { where(:active => true).order(:sorter) }
+    scope :inactive, -> { where(:active => false).order(:sorter) }
+    scope :default, -> { where(:default => true).order(:sorter) }
+    scope :not_default, -> { where(:default => false).order(:sorter) }
 
     if ActiveRecord::Base.connection.table_exists?("goldencobra_settings")
       if Goldencobra::Setting.for_key("goldencobra.widgets.recreate_cache") == "true"
