@@ -1,14 +1,13 @@
 ActiveAdmin.register Goldencobra::Articletype, :as => "Articletype" do
   menu :parent => I18n.t("active_admin.articles.parent"), :label => I18n.t('active_admin.articletypes.as'), :if => proc{can?(:update, Goldencobra::Article)}
 
-  controller.authorize_resource :class => Goldencobra::Article
   config.clear_action_items!
 
   index :download_links => proc{ Goldencobra::Setting.for_key("goldencobra.backend.index.download_links") == "true" }.call do
     selectable_column
     column :name
     column :default_template_file
-    default_actions
+    actions
   end
 
   form :html => { :enctype => "multipart/form-data" }  do |f|
@@ -34,7 +33,7 @@ ActiveAdmin.register Goldencobra::Articletype, :as => "Articletype" do
     if f.object.present? && !f.object.new_record? && File.exists?("#{::Rails.root}/app/views/articletypes/#{f.object.name.underscore.parameterize.downcase}/_edit_articletype.html.erb")
       render :partial => "articletypes/#{f.object.name.underscore.parameterize.downcase}/edit_articletype", :locals => {:f => f}
     end
-	  Rails::Application::Railties.engines.select{|a| a.engine_name.include?("goldencobra")}.each do |engine|
+	  ::Rails::Engine.subclasses.map(&:instance).select{|a| a.engine_name.include?("goldencobra")}.each do |engine|
 	  	if File.exists?("#{engine.root}/app/views/layouts/#{engine.engine_name}/_edit_articletype.html.erb")
 	  	  render :partial => "layouts/#{engine.engine_name}/edit_articletype", :locals => {:f => f, :engine => engine}
 	  	end
