@@ -3,8 +3,6 @@
 ActiveAdmin.register Goldencobra::Menue, :as => "Menue" do
   menu :parent => I18n.t('active_admin.menues.parent'), :label => I18n.t('active_admin.menues.as'), :if => proc{can?(:update, Goldencobra::Menue)}
 
-  controller.authorize_resource :class => Goldencobra::Menue
-
   filter :title, :label => I18n.t('active_admin.menues.labels.title')
   filter :target, :label => I18n.t('active_admin.menues.labels.target')
   filter :css_class, :label => I18n.t('active_admin.menues.labels.css_class')
@@ -109,7 +107,7 @@ ActiveAdmin.register Goldencobra::Menue, :as => "Menue" do
       menue.active = false
       menue.save
     end
-    flash[:notice] = I18n.t('active_admin.menues.form.column.batch_action.flash')
+    flash["notice"] = I18n.t('active_admin.menues.form.column.batch_action.flash')
     redirect_to :action => :index
   end
 
@@ -118,7 +116,7 @@ ActiveAdmin.register Goldencobra::Menue, :as => "Menue" do
       menue.active = true
       menue.save
     end
-    flash[:notice] = I18n.t('active_admin.menues.form.column.batch_action.flash1')
+    flash["notice"] = I18n.t('active_admin.menues.form.column.batch_action.flash1')
     redirect_to :action => :index
   end
 
@@ -146,7 +144,7 @@ ActiveAdmin.register Goldencobra::Menue, :as => "Menue" do
     Goldencobra::Menue.find(selection).each do |menue|
       cloned_parent = Goldencobra::Menue.create(:title => "#{I18n.t('active_admin.menues.form.batch_action.title_clone')} #{menue.title}", :target => menue.target, :css_class => menue.css_class, :active => menue.active, :parent_id => menue.parent_id, :sorter => menue.sorter, :description => menue.description, :call_to_action_name => menue.call_to_action_name, :description_title => menue.description_title, :image_id => menue.image_id)
     end
-    flash[:notice] = I18n.t('active_admin.menues.form.batch_action.flash_clone')
+    flash["notice"] = I18n.t('active_admin.menues.form.batch_action.flash_clone')
     redirect_to :action => :index
   end
 
@@ -167,7 +165,7 @@ ActiveAdmin.register Goldencobra::Menue, :as => "Menue" do
   controller do
     def show
       show! do |format|
-         format.html { redirect_to edit_admin_menue_path(@menue), :flash => flash }
+         format.html { redirect_to edit_admin_menue_path(@menue)}
       end
     end
   end
