@@ -11,7 +11,7 @@ module Goldencobra
 
     ArticleFieldOptions = {
         :global_sorting_id => %{<% f.input :global_sorting_id, :hint => I18n.t("goldencobra.article_field_hints.global_sorting_id") %>},
-        :title => %{<% f.input :title, :hint => I18n.t("goldencobra.article_field_hints.title") %>},
+        :title => '<% f.input :title, :hint => I18n.t("goldencobra.article_field_hints.title") %>',
         :subtitle => %{<% f.input :subtitle %>},
         :content => %{<% f.input :content, :input_html => { :class => "tinymce" } %>},
         :teaser => %{<% f.input :teaser, :hint => I18n.t("goldencobra.article_field_hints.teaser"), :input_html => { :rows => 5 } %>},
