@@ -6,6 +6,6 @@ module Goldencobra
 
     belongs_to :group, :class_name => Goldencobra::ArticletypeGroup, :foreign_key => :articletype_group_id
 
-    default_scope order(:sorter)
+    default_scope { order(:sorter) }
   end
 end
