@@ -8,6 +8,6 @@ module Goldencobra
     has_many :fields, :class_name => Goldencobra::ArticletypeField, :dependent => :delete_all
     accepts_nested_attributes_for :fields, :allow_destroy => true
 
-    default_scope order(:sorter)
+    default_scope { order(:sorter) }
   end
 end
