@@ -30,12 +30,8 @@ module Goldencobra
     after_update :update_cache
 
     scope :parent_ids_in_eq, lambda { |art_id| subtree_of(art_id) }
-    search_methods :parent_ids_in_eq
-
     scope :parent_ids_in, lambda { |art_id| subtree_of(art_id) } if ActiveRecord::Base.connection.table_exists?("goldencobra_settings") && Goldencobra::Setting.all.any?
-    search_methods :parent_ids_in
-
-    scope :with_values, where("value IS NOT NULL")
+    scope :with_values, -> { where("value IS NOT NULL") }
 
 
     def self.absolute_base_url
@@ -195,6 +191,25 @@ module Goldencobra
         FileUtils.touch("tmp/settings/updated_#{name}.txt")
         return Time.now
       end
+    end
+
+
+    # **************************
+    # **************************
+    # Private Methods
+    # **************************
+    # **************************
+
+    private
+
+    # Allow Scopes and Methods to search for in ransack (n.a. metasearch)
+    # @param auth_object = nil [self] "if auth_object.try(:admin?)"
+    # 
+    # @return [Array] Array of Symbols representing scopes and class methods
+    def self.ransackable_scopes(auth_object = nil)
+      [
+       :parent_ids_in
+      ]
     end
 
   end
