@@ -1,10 +1,10 @@
-require 'compass'
-require 'compass/app_integration/rails'
+# require 'compass'
+# require 'compass/app_integration/rails'
 
 Sass::Engine::DEFAULT_OPTIONS[:load_paths].tap do |load_paths|
-    load_paths << "#{Rails.root}/app/assets/stylesheets"
-    load_paths << "#{Goldencobra::Engine.root}/app/assets/stylesheets"
-    load_paths << "#{Gem.loaded_specs['compass'].full_gem_path}/frameworks/compass/stylesheets"
+  load_paths << "#{Rails.root}/app/assets/stylesheets"
+  load_paths << "#{Goldencobra::Engine.root}/app/assets/stylesheets"
+  load_paths << "#{Gem.loaded_specs['compass'].full_gem_path}/frameworks/compass/stylesheets"
 end
 
 #configuration = StringIO.new(<<-CONFIG)
