@@ -1,7 +1,7 @@
 class AddCacheToGoldencobraMenues < ActiveRecord::Migration
   def up
   	add_column :goldencobra_menues, :ancestry_depth, :integer, :default => 0
-  	Goldencobra::Menue.scoped.each do |m|
+  	Goldencobra::Menue.all.each do |m|
   		m.save
   	end
   end
