@@ -1,32 +1,5 @@
 # encoding: utf-8
 
-require "friendly_id"
-require 'ancestry'
-require 'devise'
-require 'cancan'
-require 'meta_tags'
-require 'sass'
-require 'sprockets'
-require 'sprockets/railtie'
-require 'sass-rails'
-require 'acts-as-taggable-on'
-require 'paperclip'
-require 'liquid'
-require 'geocoder'
-require "paper_trail"
-require 'sunspot_rails'
-require 'sunspot_solr'
-# require "pdfkit"
-# require 'wkhtmltopdf-binary'
-# require "wicked_pdf"
-require 'sidekiq'
-require 'sinatra'
-require 'slim'
-require 'geokit'
-require "rack/utf8_sanitizer"
-require 'simple_enum'
-require 'active_model_serializers'
-
 module Goldencobra
   class Engine < ::Rails::Engine
 
@@ -62,11 +35,15 @@ module Goldencobra
 
     end
 
-    require "#{Goldencobra::Engine.root}/app/middleware/goldencobra/handle_invalid_percent_encoding.rb"
-    # NOTE: These must be in this order relative to each other.
-    # HandleInvalidPercentEncoding just raises for encoding errors it doesn't cover,
-    # so it must run after (= be inserted before) Rack::UTF8Sanitizer.
-    config.middleware.insert 0, Goldencobra::HandleInvalidPercentEncoding
-    config.middleware.insert 0, Rack::UTF8Sanitizer  # from a gem
+    if defined? Rack::UTF8Sanitizer
+      require "#{Goldencobra::Engine.root}/app/middleware/goldencobra/handle_invalid_percent_encoding.rb"
+      # require "rack/utf8_sanitizer"
+
+      # NOTE: These must be in this order relative to each other.
+      # HandleInvalidPercentEncoding just raises for encoding errors it doesn't cover,
+      # so it must run after (= be inserted before) Rack::UTF8Sanitizer.
+      config.middleware.insert 0, Goldencobra::HandleInvalidPercentEncoding
+      config.middleware.insert 0, Rack::UTF8Sanitizer # from a gem
+    end
   end
 end
