@@ -1,6 +1,8 @@
 Dummy::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
+  config.eager_load = true
+  
   # Code is not reloaded between requests
   config.cache_classes = true
 
@@ -62,9 +64,6 @@ Dummy::Application.configure do
   # Send deprecation notices to registered listeners
   config.active_support.deprecation = :notify
 
-  # Log the query plan for queries taking more than this (works
-  # with SQLite, MySQL, and PostgreSQL)
-  # config.active_record.auto_explain_threshold_in_seconds = 0.5
   #config.middleware.use ExceptionNotifier,
   #  :sender_address => 'no_reply@goldencobra.de'
   #  :exception_recipients => %w(holger@ikusei.de, marco.metz@ikusei.de)
