@@ -1,8 +1,8 @@
 # encoding: utf-8
-
+require 'inherited_resources'
 Rails.application.routes.draw do
-  ActiveAdmin.routes(self)
   devise_for :users, ActiveAdmin::Devise.config
+  ActiveAdmin.routes(self)
   mount Goldencobra::Engine => "/"
 end
 #== Route Map
